+++ packages/orq-rc/src/models/components/evaluatorresponseragas.ts
@@ -54,8 +54,8 @@ export const EvaluatorResponseRagas$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-24T09:28:41.056Z"),
-  updated: z.string().default("2026-08-24T09:28:41.056Z"),
+  created: z.string().default("2026-08-24T14:41:02.470Z"),
+  updated: z.string().default("2026-08-24T14:41:02.470Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),