+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-24T09:28:39.484Z",
+    "2026-08-24T14:41:00.800Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {