+++ packages/orq-rc/RELEASES.md
@@ -11829,3 +11829,13 @@ Based on:
 - [typescript v4.14.0-rc.47] packages/orq-rc
 ### Releases
 - [NPM v4.14.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.47 - packages/orq-rc
+
+## 2026-08-24 14:44:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.2 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.14.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v4.14.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.48 - packages/orq-rc